+++ app/SearchApi/BcGov.Fams3.Utils/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace BcGov.Fams3.Utils
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}
